# backend/src/routes/nearbyRoutes.js
// Función "Cerca": descubrir comunidad en tu zona con ubicación ofuscada.
// Principios: (1) el servidor solo recibe la CELDA de ~2 km calculada
// en el navegador — nunca coordenadas; (2) recíproco: solo ves si compartes;
// (3) la celda caduca a los 7 días; (4) cuadrícula fija anti-triangulación.
const express = require('express');
const rateLimit = require('express-rate-limit');
const router = express.Router();

const prisma = require('../lib/prisma');
const { requireAuth, requireNotSuspended } = require('../middlewares/requireAuth');
const { isValidCell, centroid, neighborsGrid, cellDistanceKm } = require('../lib/geogrid');
const { blockedWith, isBlockedBetween } = require('../lib/blocks');

const CELL_TTL_DAYS = 7;
const GRID_RINGS = 5; // 11×11 celdas de ~2 km ≈ radio efectivo ~11 km
const POKE_COOLDOWN_HOURS = 12;

const participantSelect = { id: true, name: true, displayName: true, avatar: true, acct: true };

function cutoffDate() {
  return new Date(Date.now() - CELL_TTL_DAYS * 24 * 60 * 60 * 1000);
}

// Anti-scraping: el mapa no se consulta en ráfaga
const nearbyLimiter = rateLimit({
  windowMs: 5 * 60 * 1000,
  limit: 60,
  standardHeaders: 'draft-7',
  legacyHeaders: false,
  message: { error: 'Demasiadas consultas al mapa. Espera unos minutos.' }
});

function bandLabel(km, sameCell) {
  if (sameCell) return 'En tu zona';
  const rounded = Math.max(2, Math.round(km / 2) * 2);
  return `A ~${rounded} km`;
}

// Una celda vigente = existe, es del formato actual (las del geohash viejo se
// descartan) y no ha caducado
function hasActiveCell(user) {
  return Boolean(user?.nearbyCell && isValidCell(user.nearbyCell) && user.nearbyUpdatedAt >= cutoffDate());
}

// GET /api/nearby/location — mi estado de compartir (celda propia o null)
router.get('/location', requireAuth, async (req, res) => {
  try {
    const me = await prisma.user.findUnique({
      where: { id: req.user.id },
      select: { nearbyCell: true, nearbyUpdatedAt: true }
    });
    const active = hasActiveCell(me);
    res.json({ sharing: active, cell: active ? me.nearbyCell : null, updatedAt: active ? me.nearbyUpdatedAt : null });
  } catch (e) {
    console.error('Error al consultar estado de Cerca:', e);
    res.status(500).json({ error: 'Error al consultar tu estado' });
  }
});

// PUT /api/nearby/location { cell } — activar/actualizar mi zona
router.put('/location', requireAuth, async (req, res) => {
  // Defensa explícita: si un cliente manda coordenadas reales, se rechaza y no se registra nada
  const forbidden = ['lat', 'lng', 'lon', 'latitude', 'longitude', 'coords', 'accuracy'];
  if (forbidden.some(k => k in (req.body || {}))) {
    return res.status(400).json({ error: 'Este endpoint solo acepta la celda ofuscada, nunca coordenadas' });
  }
  const cell = typeof req.body.cell === 'string' ? req.body.cell.trim() : '';
  if (!isValidCell(cell)) {
    return res.status(400).json({ error: 'Celda inválida' });
  }
  try {
    await prisma.user.update({
      where: { id: req.user.id },
      data: { nearbyCell: cell, nearbyUpdatedAt: new Date() }
    });
    res.json({ sharing: true, cell, updatedAt: new Date() });
  } catch (e) {
    console.error('Error al actualizar zona:', e);
    res.status(500).json({ error: 'Error al actualizar tu zona' });
  }
});

// DELETE /api/nearby/location — dejar de compartir (borra la celda)
router.delete('/location', requireAuth, async (req, res) => {
  try {
    await prisma.user.update({
      where: { id: req.user.id },
      data: { nearbyCell: null, nearbyUpdatedAt: null }
    });
    res.json({ sharing: false });
  } catch (e) {
    console.error('Error al dejar de compartir zona:', e);
    res.status(500).json({ error: 'Error al dejar de compartir' });
  }
});

// GET /api/nearby — personas en mi zona y alrededores (recíproco)
router.get('/', requireAuth, nearbyLimiter, async (req, res) => {
  try {
    const me = await prisma.user.findUnique({
      where: { id: req.user.id },
      select: { nearbyCell: true, nearbyUpdatedAt: true }
    });
    if (!hasActiveCell(me)) {
      return res.status(403).json({ error: 'Comparte tu zona para ver quién anda cerca (es recíproco)' });
    }

    const cells = neighborsGrid(me.nearbyCell, GRID_RINGS);
    // Quien está bloqueado (en cualquier dirección) desaparece del mapa para ambos
    const hidden = await blockedWith(req.user.id);
    const users = await prisma.user.findMany({
      where: {
        id: { notIn: [req.user.id, ...hidden] },
        nearbyCell: { in: cells },
        nearbyUpdatedAt: { gte: cutoffDate() }
      },
      select: { ...participantSelect, nearbyCell: true }
    });

    const people = users
      .map(u => {
        const km = cellDistanceKm(me.nearbyCell, u.nearbyCell);
        const { nearbyCell, ...pub } = u;
        return { ...pub, cell: nearbyCell, distanceKm: Math.round(km), band: bandLabel(km, nearbyCell === me.nearbyCell) };
      })
      .sort((a, b) => a.distanceKm - b.distanceKm);

    // Zonas agregadas para el mapa (centroide + conteo por celda)
    const zoneMap = new Map();
    for (const p of people) {
      if (!zoneMap.has(p.cell)) {
        const c = centroid(p.cell);
        zoneMap.set(p.cell, { cell: p.cell, lat: c.lat, lon: c.lon, count: 0 });
      }
      zoneMap.get(p.cell).count += 1;
    }

    const myCentroid = centroid(me.nearbyCell);
    res.json({
      myZone: { cell: me.nearbyCell, lat: myCentroid.lat, lon: myCentroid.lon },
      people,
      zones: [...zoneMap.values()]
    });
  } catch (e) {
    console.error('Error al consultar Cerca:', e);
    res.status(500).json({ error: 'Error al consultar la zona' });
  }
});

// POST /api/nearby/poke { userId } — mandar un toque 👋 (llega como notificación)
//
// El toque es parte de "Cerca" y hereda sus dos reglas (HU-SEG-004): solo lo manda
// quien comparte zona, y solo llega a quien está dentro de su cuadrícula. Sin estas
// comprobaciones el endpoint era un "ping a cualquier userId": como los ids son
// enteros consecutivos, bastaba recorrerlos para notificar a toda la base.
router.post('/poke', requireAuth, requireNotSuspended, nearbyLimiter, async (req, res) => {
  const targetId = Number(req.body.userId);
  if (!targetId) return res.status(400).json({ error: 'userId requerido' });
  if (targetId === req.user.id) return res.status(400).json({ error: 'No puedes mandarte un toque a ti' });
  try {
    // 1. Reciprocidad: sin zona activa propia no se manda nada
    const me = await prisma.user.findUnique({
      where: { id: req.user.id },
      select: { nearbyCell: true, nearbyUpdatedAt: true }
    });
    if (!hasActiveCell(me)) {
      return res.status(403).json({ error: 'Comparte tu zona para mandar un toque (es recíproco)' });
    }

    const target = await prisma.user.findUnique({
      where: { id: targetId },
      select: { id: true, nearbyCell: true, nearbyUpdatedAt: true }
    });
    // 2. Cercanía: el destino debe estar compartiendo y caer dentro de mi cuadrícula.
    // Un destino inexistente, lejano o que no comparte responden igual (404): el
    // resultado no debe servir para deducir dónde está ni si existe.
    const reachable = target
      && hasActiveCell(target)
      && neighborsGrid(me.nearbyCell, GRID_RINGS).includes(target.nearbyCell);
    if (!reachable) return res.status(404).json({ error: 'Esa persona no está cerca' });

    // 3. Bloqueo en cualquier dirección: mismo 404, sin revelar el motivo
    if (await isBlockedBetween(req.user.id, targetId)) {
      return res.status(404).json({ error: 'Esa persona no está cerca' });
    }

    // 4. Anti-spam: un toque por persona cada POKE_COOLDOWN_HOURS
    const since = new Date(Date.now() - POKE_COOLDOWN_HOURS * 60 * 60 * 1000);
    const recent = await prisma.notification.findFirst({
      where: { type: 'POKE', actorId: req.user.id, recipientId: targetId, createdAt: { gte: since } },
      select: { id: true }
    });
    if (recent) {
      return res.status(429).json({ error: 'Ya le mandaste un toque hace poco — dale chance de responder 🌿' });
    }

    await prisma.notification.create({
      data: { type: 'POKE', recipientId: targetId, actorId: req.user.id }
    });
    res.json({ ok: true });
  } catch (e) {
    console.error('Error al mandar toque:', e);
    res.status(500).json({ error: 'No se pudo mandar el toque' });
  }
});

module.exports = router;