# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeedTown API",
    "version": "2.4.0",
    "description": "API de WeedTown, red social de la comunidad cannábica de México. Identidad federada con Mastodon (OAuth 2.0) y sesión mediante JWT propio en el header Authorization. Seguridad: headers via helmet, CORS restringido al frontend, rate limiting (300 peticiones/15 min general; 20/15 min en el flujo OAuth), payload JSON limitado a 100 kB e imágenes a 5 MB. El perfil público (GET /api/profile/{id}) no expone PII (email, teléfono, nombre real, edad, nacimiento ni género). Las rutas de mercado y admin son stubs pendientes de implementación."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por el backend tras el flujo OAuth de Mastodon (expira en 7 días)."
      }
    },
...
  },
...
  "paths": {
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notificaciones"
        ],
        "summary": "Últimas notificaciones del usuario (REPLY_POST, REPLY_COMMENT, NEW_SUBFORUM_POST)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ notifications[], unread }"
          }
        }
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "tags": [
          "Notificaciones"
        ],
        "summary": "Contador de no leídas (para el badge de la campana)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "{ count }"
          }
        }
      }
    },
    "/api/notifications/read-all": {
      "post": {
        "tags": [
          "Notificaciones"
        ],
        "summary": "Marcar todas como leídas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "{ ok: true }"
          }
        }
      }
    },
...
    "/api/nearby/location": {
      "get": {
        "tags": [
          "Cerca"
        ],
        "summary": "Mi estado de compartir zona",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "sharing, cell (geohash-5) y updatedAt"
          }
        }
      },
      "put": {
        "tags": [
          "Cerca"
        ],
        "summary": "Activar/actualizar mi zona (celda de cuadrícula ~2 km ofuscada en el cliente)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Solo acepta la celda ({latIdx}_{lonIdx}, cuadrícula fija de 0.02° ≈ 2 km); cualquier payload con lat/lng/coords se rechaza con 400. La celda caduca a los 7 días.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cell"
                ],
                "properties": {
                  "cell": {
                    "type": "string",
                    "pattern": "^d{1,4}_d{1,5}$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zona actualizada"
          },
          "400": {
            "description": "Celda inválida o intento de enviar coordenadas"
          }
        }
      },
      "delete": {
        "tags": [
          "Cerca"
        ],
        "summary": "Dejar de compartir (borra la celda)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "sharing=false"
          }
        }
      }
    },
    "/api/nearby": {
      "get": {
        "tags": [
          "Cerca"
        ],
        "summary": "Personas y zonas cercanas (recíproco: exige compartir zona vigente)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Busca en una cuadrícula 11×11 de celdas de ~2 km (~11 km de radio). Devuelve perfiles públicos sin PII con banda de distancia redondeada a 2 km, y zonas agregadas con conteo para el mapa. Rate limit propio. Las personas con un bloqueo de por medio no aparecen en la lista ni cuentan en las zonas del mapa.",
        "responses": {
          "200": {
            "description": "myZone, people[] y zones[]"
          },
          "403": {
            "description": "No compartes tu zona (o caducó)"
          },
          "429": {
            "description": "Rate limit"
          }
        }
      }
    },
    "/api/nearby/poke": {
      "post": {
        "tags": [
          "Cerca"
        ],
        "summary": "Mandar un toque 👋 (llega como notificación in-app)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Invitación ligera a interactuar. Hereda las reglas de Cerca: hay que compartir zona para mandarlo (recíproco) y la persona destinataria debe estar dentro de la cuadrícula de ~11 km. Anti-spam: máximo un toque por persona cada 12 horas, más el rate limit propio de Cerca (60 por 5 min).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Toque enviado"
          },
          "400": {
            "description": "userId inválido o toque a ti mismo"
          },
          "403": {
            "description": "No compartes tu zona (el toque es recíproco)"
          },
          "404": {
            "description": "Esa persona no está cerca. Misma respuesta si no existe, si no comparte zona, si está fuera de la cuadrícula o si hay un bloqueo de por medio: el resultado no debe permitir deducir nada."
          },
          "429": {
            "description": "Cooldown de 12 h activo, o rate limit del mapa"
          }
        }
      }
    },
...
  }
}